# public/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentUtils API v2",
    "version": "2.0.0",
    "description": "Multi-tenant, agent-native infrastructure: KV store, audit log, dead-letter queue, scheduler, human-in-the-loop checkpoints, and image upload. Tenant-isolated, callback-signed, idempotent.\n\n## Base URL\nhttps://www.agent-utils.com/v1\n\n## Authentication\nServer-derived identity from key prefix:\n- `agutil_adm_\u2026` \u2014 admin key (tenant scope). Header: `x-admin-key`.\n- `agutil_agt_\u2026` \u2014 agent key (tenant + agent scope). Headers: `x-agent-id` + `x-api-key`.\n- `agutil_apr_\u2026` \u2014 approval-proxy key (HitL approve/reject only). Header: `X-Approval-Key`.\n\n`tenant_id` is always server-derived \u2014 never accepted from the client.\n\n## Standard headers\n- `Idempotency-Key` \u2014 replay-safe on all `POST` creation endpoints.\n- All responses include `X-Request-Id`.\n\n## Error envelope\n`{ \"error\": { \"code\", \"message\", \"details\", \"request_id\" } }`"
  },
  "servers": [
    {
      "url": "https://www.agent-utils.com/v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Admin key `agutil_adm_\u2026`"
      },
      "agentKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Agent key `agutil_agt_\u2026` (with `x-agent-id`)"
      },
      "approvalKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Approval-Key",
        "description": "Approval-proxy key `agutil_apr_\u2026`"
      },
      "connectionCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "au_conn_\u2026",
        "description": "Connection credential returned once when a pairing code is redeemed"
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET bearer for the cron tick"
      }
    },
...
    "schemas": {
...
      "AgentCreate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{2,31}$"
          },
          "description": {
            "type": "string"
          },
          "callback_base_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/agents": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Register agent (returns one-time agent key)",
        "security": [
          {
            "adminKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created \u2014 one-time `agutil_agt_\u2026` key"
          },
          "409": {
            "description": "AGENT_ID_TAKEN"
          },
          "402": {
            "description": "QUOTA_EXCEEDED (agents)"
          }
        }
      }
    },
...
  }
}